Frandom support
Frandom is a Linux kernel random number generator,
which is 10-50 times faster than what you get from Linux'
built-in /dev/urandom. And it uses very little (/dev/frandom)
or none (/dev/erandom) of the kernel's entropy pool,
so it is very useful for applications that require a
handy source for lots of random data.

The test-results i get are quite good